Initial commit of module LaPermanence.py
Initial commit of module LaPermanence.py`

Note: there is no documentation in this initial commit.

Introduce three classes:

Class Location
This is a glorified dictionary. Contains information on locationso of
coworking spaces.

Class DataManager
This is a convenience class, which loads the data from the csv file
and generates pivot tables by displaying the number of available seats
at a selected location as a calendar.  For examples, each column
contains the data for one day in the dataset, and all days in the
dataset fill the columns; or each column represents the average for
each day of the week, and the 7 columns in the pivot table represent
the days of the week; similarly each column may represent the data for
one week in the dataset.  The "mode" refers to the frequency
represented in each column: "Date" for one day in the dataset,
"DayOfWeek" for one day of week, etc.

These pivot tables are saved as objects of the class View.

Class View
This class consists essentially of pivot tables along with the
corresponding location and the mode.

On branch master
Changes to be committed:
	modified:   ../.gitignore
	new file:   LaPermanence.py
	modified:   ../scripts/backup.py
	modified:   ../scripts/scraping.py

Changes not staged for commit:
	modified:   ../notebooks/la-permanence-decision-trees.ipynb
	modified:   ../notebooks/la-permanence-random-forests-feature-importance.ipynb
	modified:   ../notebooks/la-permanence-random-forests-parameters.ipynb
	modified:   ../notebooks/la-permanence-understanding.ipynb