Include scraping, automate interaction with kaggle and github, improve logging
SCRAPING
The script for scraping now essentially reduces to a function from the
module.

KAGGLE & GIT/GITHUB
Add functionalities:
- to download dataset from kaggle
- upload dataset to kaggle
- git add/commmit/push figures to github page (not this repo)

The classes DataManager and View have been streamlined, with some
change in the constructors.

The logging has been streamlined.

The three main scripts are:
- scraping.py: collect data;
- backup.py: backup hourly, daily, weekly (still needs to be fixed);
- update.py: download dataset from kaggle, combine with local dataset,
generate new figures for github page, and push figures to repo.

Each script is scheduled to run with in a cron job, see corresponding
cron-XXX.bash script.