Add script correctinf format of csv file
One location has been added to the network of La Permanence.  As a
result, the recent records contain one more column than all previous.
The script add-field-better.py includes the name of the new location
"marcadet" on the first line, and adds a comma "," at rows missing
this field.

On branch master
Changes to be committed:
	new file:   tmp/add-field-better.py

Changes not staged for commit:
	modified:   .gitignore
	modified:   notebooks/la-permanence-decision-trees.ipynb
	modified:   notebooks/la-permanence-random-forests-feature-importance.ipynb
	modified:   notebooks/la-permanence-random-forests-parameters.ipynb
	modified:   notebooks/la-permanence-understanding.ipynb

Untracked files:
	scripts/environment.yml